Insert and lookup for trailers.
